Fixed : to each page the forbidden

+++ sfapp/src/Controller/RoomsController.php
@@ -60,12 +60,14 @@ class RoomsController extends AbstractController
     }
 
    
-    #[IsGranted("ROLE_ADMIN")]
     #[Route('/rooms/add', name: 'app_room_add')]
     public function add(Request $request, EntityManagerInterface $entityManager): Response
     {
         $room = new Room();
         $form = $this->createForm(RoomFormType::class, $room);
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_error_403');
+        }
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -84,10 +86,12 @@ class RoomsController extends AbstractController
     }
 
     
-    #[IsGranted("ROLE_ADMIN")]
     #[Route('/rooms/{id}', name: 'app_room_delete', methods: ['POST'])]
     public function delete(Request $request, Room $room, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_error_403');
+        }
         $entityManager->remove($room);
         $entityManager->flush();
 
@@ -98,10 +102,12 @@ class RoomsController extends AbstractController
     }
 
     
-    #[IsGranted("ROLE_ADMIN")]
     #[Route('/rooms/{id}/edit', name: 'app_room_edit')]
     public function edit(Room $room, Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_error_403');
+        }
         $form = $this->createForm(RoomFormType::class, $room);
 
         $form->handleRequest($request);
